requirements.txt
httpx
py-clob-client
python-dotenv